# Add image ufunc API