--- a/src/sknnr_spatial/image.py
+++ b/src/sknnr_spatial/image.py
@@ -204,8 +204,7 @@ class DataArrayImage(Image):
 
     @property
     def band_names(self) -> NDArray:
-        band_dim_name = self.image.dims[self.band_dim]
-        return self.image[band_dim_name].values
+        return self.image[self.band_dim_name].values
 
     def _validate_nodata_vals(self, nodata_vals: NoDataType) -> NDArray | None:
         """