--- a/src/sknnr_spatial/image.py
+++ b/src/sknnr_spatial/image.py
@@ -263,7 +263,7 @@ class DataArrayImage(Image):
         n_outputs = len(output_dims)
         # Fall back to float output if unknown
         output_dtypes = output_dtypes or [np.float32] * n_outputs
-        # If output sizes are not provided, assume a single output coordinate
+        # Default to one output band per input band
         output_sizes = output_sizes or {"variable": len(self.band_names)}
         # Default to sequential coordinates for each output dimension, if not provided
         output_coords = output_coords or {