… src/sknnr_spatial/image/_base.py
… src/sknnr_spatial/image/_dask_backed.py
… src/sknnr_spatial/image/dataarray.py
… src/sknnr_spatial/image/dataset.py
… src/sknnr_spatial/image/ndarray.py
… tests/test_estimator.py
… tests/test_image.py
… tests/test_images.py
… tests/test_predict.py
… tests/test_sknnr.py
… tests/test_wrapper.py