This will eventually close #27 by adding a more flexible API for running ufuncs across image types and migrating our `sklearn` methods over to that. This is going to be a substantial refactor of the package internals, so I'm opening this as a draft for now to allow for performance testing and discuss details. At this point, most of the tests are failing and the outputs aren't necessarily consistent with the previous API, but *most* of the features are in place.

### TODO
- [x] Implement `Image.apply_ufunc_across_bands`
- [x] Migrate `predict` and `kneighbors` to use that
- [ ] Implement `xr.Dataset` outputs
- [ ] Probably refactor `Image` into subclasses based on image types, so that they can handle things like NoData parsing separately. Should be able to re-use most of the existing code for that functionality.
- [ ] Get NoData re-implemented
- [ ] Get feature name checks re-implemented
- [ ] Output `xr.Dataset` from `xr.Dataset` inputs (currently they're `xr.DataArray`)
- [ ] If possible, simplify the `apply_ufunc_across_bands` arguments
- [ ] Get all tests working again (rewrite as needed)

### Performance

Here's the test code I ran to compare `predict` and `kneighbors` compute times for the large SWO rasters:

```python
import timeit

from sknnr import GNNRegressor
from threadpoolctl import threadpool_limits

from sknnr_spatial import wrap
from sknnr_spatial.datasets import load_swo_ecoplot

if __name__ == "__main__":
    X_img, X, y = load_swo_ecoplot(as_dataset=True, large_rasters=True)
    est = wrap(GNNRegressor()).fit(X, y)

    with threadpool_limits(limits=1, user_api="blas"):
        print(f"predict: {timeit.timeit(lambda: est.predict(X_img).compute(), number=1):.03f}s")
        print(f"kneighbors: {timeit.timeit(lambda: [x.compute() for x in est.kneighbors(X_img)], number=1):.03f}s")
```

The inclusion of `threadpool_limits` fixed a threading issue I was having with both branches, discussed in #19.

The results show that the new `ufunc` branch implementation is **not any slower** than the original implementation for this raster size: 

| Branch  | `predict` | `kneighbors` |
|---------|-----------|--------------|
| `main`  | 16.812s   | 28.781s      |
| `ufunc` | 15.330s   | 28.314s      |

@grovduck, if/when you get a chance (no rush), could you try running your regional GNN map with this branch to make sure that performance still holds up with a real-world dataset?