@grovduck 2024-07-01
> @grovduck, if/when you get a chance (no rush), could you try running your regional GNN map with this branch to make sure that performance still holds up with a real-world dataset?

@aazuspan, for testing, do you have specific schedules / numbers of threads you'd like me to test?  I'm wondering if I'm doing something wrong because even if I run an area twice as large as the sample SWO large dataset, it's taking much more than twice as long.  The differences between the run you show above and the MR224 run are:

- I'm using `xarray.Dataset.rio.isel_window` to subset the rasters based on bounds from full OR/CA/WA rasters
- The X, y dataframes are slightly bigger (n=4,314 vs n=3,005)
- The test rasters I'm using are 8200 by 2048

I'll send you a gist of the code I'm using as well as some pointers to the data.