@grovduck 2024-07-15
@aazuspan, I feel like I sidetracked us with the performance discussion here.  I'm happy to start reviewing this PR if it makes sense at this stage.  I see there are a few outstanding tasks, so I can wait as well.