@aazuspan 2024-07-15
@grovduck, no worries, I actually got derailed by being sick for the last couple weeks (flu and cold back-to-back!) and figured it was best not to mix complicated refactoring with cold medicine! Hoping to start making some headway now, though. I'd probably recommend holding off on the review since there might be a lot of refactoring left to do still, but thanks!