@grovduck 2024-07-15
> I actually got derailed by being sick for the last couple weeks (flu and cold back-to-back!) and figured it was best not to mix complicated refactoring with cold medicine!

I would swear that you're more productive on cold meds than I am when functioning at full steam.  Glad to hear you're feeling better.