@aazuspan 2024-07-18
@grovduck, I think this is finally ready for a review whenever you've got some down time. I edited the original post to try and walk through the major design changes. I'm open to any thoughts you might have on design, naming, etc!

Also, since parallelizing the calculation of weighted attributes from a neighbor grid was the motivation for this, maybe we should try to implement that using this branch and see if we hit any snags?