@aazuspan 2024-07-20
> I did a bit of rubber-stamping on my review of all the tests, but I did run tests and they succeeded for me locally.

Sorry, I created a lot of diff noise in the tests by reorganizing and renaming files! I think the real changes were pretty minimal - mostly just swapping the preprocessor system for `Image.from_image` and dropping a few image tests like `test_flatten` and `test_flatten_is_reversible` since those are indirectly covered.

> I will spend some quality time going over your example of using this for weighted neighbor attribute calculation to get more familiar with the API.

Sounds good, thanks!!