@grovduck 2024-07-22
@aazuspan, I think all my questions/issues are resolved (either by what you've already done or by putting into separate PRs) other than the `output_sizes` question on `apply_ufunc_across_bands`.