@aazuspan 2024-07-23
Thanks for the review @grovduck! With the default ufunc params removed, it looks like everything's resolved or tracked in separate PRs now, so I'm going to go ahead and merge.